✨ feat: update TreeViewer component
